requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
biopython==1.84
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dm-tree==0.1.8
docker-pycreds==0.4.0
exceptiongroup==1.2.2
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
GPUtil==1.4.0
hydra-core==1.3.2
idna==3.10
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
joblib==1.4.2
lightning-utilities==0.11.7
MarkupSafe==2.1.5
mdtraj==1.10.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
omegaconf==2.3.0
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
protobuf==5.28.2
psutil==6.0.0
pyparsing==3.2.0
pytest==8.3.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.13.1
sentry-sdk==2.15.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.13.3
threadpoolctl==3.5.0
tomli==2.0.2
torch==2.4.1
torch-scatter @ https://data.pyg.org/whl/torch-2.4.1+cu124.html
torchmetrics==1.4.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.18.3
yarl==1.13.1
